Move dependencyVersionMap.ts to its own file
